refactor(desktop): move hyphen-list guard to markdown-fences lib
Review follow-ups on the rich-copy PR: relocate
protectComposerHyphenListItems from the 900-line ThreadMarkdown
component module to lib/markdown-fences.ts beside its sibling source
repair, so markdown-clipboard-html no longer imports a component file
for one string helper. Also add a regression test pinning code-block
copies to plain text — the rich clipboard flavor is message-copy only.
