+++ apps/desktop/src/renderer/src/features/thread-detail/ThreadMarkdown.tsx
@@ -26,7 +26,10 @@ import { AppIcon } from "../../components/AppIcon";
 import { CloseIcon, CopyIcon, FolderIcon, PopoutIcon } from "../../icons";
 import type { DesktopApi } from "../../lib/desktop-api";
 import { copyText } from "../../lib/copy-text";
-import { repairNestedLanguageFences } from "../../lib/markdown-fences";
+import {
+  protectComposerHyphenListItems,
+  repairNestedLanguageFences,
+} from "../../lib/markdown-fences";
 import { useMarkdownMathRuntime } from "../../lib/markdown-rendering-options";
 import {
   resolveThreadHref,
@@ -57,7 +60,11 @@ type ThreadMarkdownProps = {
   className?: string;
   desktopApi?: Pick<
     DesktopApi,
-    "copyText" | "openApplication" | "openMarkdownFileViewer" | "readMarkdownFile"
+    | "copyText"
+    | "copyRichText"
+    | "openApplication"
+    | "openMarkdownFileViewer"
+    | "readMarkdownFile"
   >;
   fileViewerContext?: MarkdownFileViewerContext;
   imageParts?: AppServerThreadImagePart[];
@@ -633,7 +640,11 @@ function MarkdownDocumentModal(props: {
   applications?: DesktopApplicationsSnapshot;
   desktopApi?: Pick<
     DesktopApi,
-    "copyText" | "openApplication" | "openMarkdownFileViewer" | "readMarkdownFile"
+    | "copyText"
+    | "copyRichText"
+    | "openApplication"
+    | "openMarkdownFileViewer"
+    | "readMarkdownFile"
   >;
   editorApplication?: EditorApplication;
   fileViewerContext?: MarkdownFileViewerContext;
@@ -862,43 +873,6 @@ function MarkdownDocumentModal(props: {
     </div>,
     document.body,
   );
-}
-
-export function protectComposerHyphenListItems(markdown: string): string {
-  const lines = markdown.split("\n");
-  let fence: { marker: "`" | "~"; length: number } | undefined;
-
-  return lines
-    .map((line) => {
-      const fenceLine = line.match(/^\s{0,3}(`{3,}|~{3,})(.*)$/);
-      if (fenceLine) {
-        const sequence = fenceLine[1] ?? "";
-        const marker = sequence[0] as "`" | "~";
-        const trailing = fenceLine[2] ?? "";
-        if (!fence) {
-          fence = { marker, length: sequence.length };
-        } else if (
-          marker === fence.marker &&
-          sequence.length >= fence.length &&
-          trailing.trim() === ""
-        ) {
-          fence = undefined;
-        }
-        return line;
-      }
-
-      if (fence) {
-        return line;
-      }
-
-      const hyphenOnlyListItem = line.match(/^(\s{0,3})-\s+(-+)(\s*)$/);
-      if (!hyphenOnlyListItem || (hyphenOnlyListItem[2]?.length ?? 0) < 2) {
-        return line;
-      }
-
-      return `${hyphenOnlyListItem[1]}- \\${hyphenOnlyListItem[2]}${hyphenOnlyListItem[3]}`;
-    })
-    .join("\n");
 }
 
 const normalizeMarkdownUrl: UrlTransform = (url) => {
